pyproject.toml
[tool.poetry]
name = "vpy"
version = "0.1.0"
description = "Ventanita for Paraguay"
authors = ["Gurupratap Matharu <gurupratap.matharu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dotenv = "^0.9.9"
gunicorn = "^23.0.0"
lxml = "^5.4.0"
wagtail = "^7.2.1"
wagtail-localize = "^1.12.2"
django-allauth = {extras = ["socialaccount"], version = "^65.13.1"}


[tool.poetry.group.dev.dependencies]
djlint = "^1.36.4"
isort = "^6.0.1"
ruff = "^0.9.8"
ipython = "^9.3.0"
coverage = "^7.8.2"
django-extensions = "^4.1"
django-debug-toolbar = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
exclude = ["*migrations*", "static"]

[tool.ruff.lint.isort]
combine-as-imports=true
known-first-party=["blog"]
known-third-party=["wagtail", "rest_framework"]
lines-after-imports=2
section-order=["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
"django"=["django"]

[tool.djlint]
blank_line_before_tag="load,extends"
blank_line_after_tag="load,extends,endblock"
close_void_tags=true
custom_blocks="toc,example"
custom_html="mjml"
exclude=".venv,venv,.tox,.eggs,.git"
extend_exclude=".custom"
format_attribute_template_tags=true
format_css=true
format_js=true
ignore_blocks="raw,example"
indent=2
max_attribute_length=120
max_line_length=120
profile="django"
# preserve_blank_lines=true
use_gitignore=true

[tool.djlint.css]
indent_size=5

[tool.djlint.js]
indent_size=5
